Move content of TodoApp to TodoList, delete TodoApp, change /tasks route render to TodoList in App
