+++ src/App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Switch } from 'react-router-dom';
 import Layout from './components/layout/Layout';
-import TodoApp from './components/TodoApp';
+import TodoList from './components/TodoList';
 
 function App() {
   return (
@@ -9,7 +9,7 @@ function App() {
       <Switch >
         <Route exact path="/" render={() => <h1>Timer</h1>} />
         <Route exact path="/settings" render={() => <h1>Timer Settings</h1>} />
-        <Route exact path="/tasks" render={() => <TodoApp />} />
+        <Route exact path="/tasks" render={() => <TodoList />} />
       </Switch>
     </Layout>
   )
